feat: add api for price search
