+++ bookfinder/src/lib/types.ts
@@ -1,0 +1,84 @@
+interface VolumeInfo {
+  title: string;
+  authors: string[];
+  publisher: string;
+  publishedDate: string;
+  description: string;
+  industryIdentifiers: {
+    type: string;
+    identifier: string;
+  }[];
+  readingModes: {
+    text: boolean;
+    image: boolean;
+  };
+  pageCount: number;
+  printType: string;
+  categories: string[];
+  averageRating: number;
+  ratingsCount: number;
+  maturityRating: string;
+  allowAnonLogging: boolean;
+  contentVersion: string;
+  panelizationSummary: {
+    containsEpubBubbles: boolean;
+    containsImageBubbles: boolean;
+  };
+  imageLinks: {
+    smallThumbnail: string;
+    thumbnail: string;
+  };
+  language: string;
+  previewLink: string;
+  infoLink: string;
+  canonicalVolumeLink: string;
+}
+
+interface SaleInfo {
+  country: string;
+  saleability: string;
+  isEbook: boolean;
+  listPrice?: {
+    amount: number;
+    currencyCode: string;
+  };
+  retailPrice?: {
+    amount: number;
+    currencyCode: string;
+  };
+  buyLink?: string;
+}
+
+interface AccessInfo {
+  country: string;
+  viewability: string;
+  embeddable: boolean;
+  publicDomain: boolean;
+  textToSpeechPermission: string;
+  epub: {
+    isAvailable: boolean;
+    acsTokenLink: string;
+  };
+  pdf: {
+    isAvailable: boolean;
+    acsTokenLink: string;
+  };
+  webReaderLink: string;
+  accessViewStatus: string;
+  quoteSharingAllowed: boolean;
+}
+
+interface SearchInfo {
+  textSnippet: string;
+}
+
+export interface GoogleBookVolume {
+  kind: string;
+  id: string;
+  etag: string;
+  selfLink: string;
+  volumeInfo: VolumeInfo;
+  saleInfo: SaleInfo;
+  accessInfo: AccessInfo;
+  searchInfo: SearchInfo;
+}
